daemon: extract event/grpc.rs from event/mod.rs
Move all gRPC-related code out of the 12,169-line event/mod.rs into a
dedicated event/grpc.rs submodule:

- session_state_to_api helper
- From<&PeerView>/TryFrom<&api::Peer>/From<api::PeerGroup> conversions
- parse_gr_api (gRPC-only GR config parser)
- peer_group_to_api
- GrpcService struct and impl GoBgpService for GrpcService (~2,100 lines)
- From<&RpkiState>/From<&Global> API conversions

Functions called from both gRPC and non-gRPC paths (parse_afi_safis_yaml,
parse_gr_yaml, add_policy_assignment, create_listen_socket) remain in
mod.rs and are exposed as pub(super) for grpc.rs to call.

Shared internal types that grpc.rs accesses are promoted to pub(super)
so the event submodule tree can use them without leaking to the wider
crate. No logic changes; all 937 tests pass unchanged.

mod.rs: 12,183 lines -> 9,507 lines
grpc.rs: 2,687 lines (new)
